requirements.txt
youtube-search-python
pytubefix
pytchat
python-vlc